# Optimize updating key by storing lease in lessor