@nekto0n 2016-10-11
There is another opportunity to avoid allocating new string to construct `LeaseItem` to perform lookup in `itemMap` (go optimizes  string construction away in `itemMap[string(key)])` if `key` is `[]bytes`), but I am not sure if it is okay to bring bytes to `Lessor`.