@xiang90 2016-10-12
Can we figure out why 

```
BenchmarkWatchableStoreSyncedCancel-4       0.82          1187249426     +144786515265.85%
```

before get this merged?