@nekto0n 2016-10-12
@xiang90 I think `BenchmarkWatchableStoreSyncedCancel` is flaky. These I my runs one after another with this patch applied:
```
BenchmarkWatchableStoreSyncedCancel-4   	       1	1158563857 ns/op	    2352 B/op	      37 allocs/op
BenchmarkWatchableStoreSyncedCancel-4   	       1	1027656738 ns/op	    2048 B/op	      32 allocs/op
BenchmarkWatchableStoreSyncedCancel-4   	      50	  21017376 ns/op	     340 B/op	       5 allocs/op
```
And these are runs on `master`:
```
BenchmarkWatchableStoreSyncedCancel-4   	      10	 112671224 ns/op	    1099 B/op	      17 allocs/op
BenchmarkWatchableStoreSyncedCancel-4   	10000000	       100 ns/op	       0 B/op	       0 allocs/op
BenchmarkWatchableStoreSyncedCancel-4   	100000000	        10.0 ns/op	       0 B/op	       0 allocs/op
```
I've never seen such inconsistent things.