@nekto0n 2016-10-12
Here cpuprofile from run which looks like this
```
BenchmarkWatchableStoreSyncedCancel-4   	   20000	     51045 ns/op	       1 B/op	       0 allocs/op
```

```
(pprof) top
58.48s of 81.18s total (72.04%)
Dropped 154 nodes (cum <= 0.41s)
Showing top 10 nodes out of 85 (cum >= 7.63s)
      flat  flat%   sum%        cum   cum%
    13.77s 16.96% 16.96%     24.67s 30.39%  runtime.scanobject
    12.04s 14.83% 31.79%     12.47s 15.36%  runtime.mapaccess2_fast64
     6.93s  8.54% 40.33%     10.28s 12.66%  runtime.evacuate
     6.57s  8.09% 48.42%     21.96s 27.05%  runtime.mapassign1
     6.14s  7.56% 55.99%      6.14s  7.56%  runtime.heapBitsForObject
     4.44s  5.47% 61.46%         5s  6.16%  runtime.greyobject
     3.33s  4.10% 65.56%      4.01s  4.94%  runtime.mapdelete
     1.82s  2.24% 67.80%      1.82s  2.24%  runtime.aeshash64
     1.73s  2.13% 69.93%      1.73s  2.13%  runtime.procyield
     1.71s  2.11% 72.04%      7.63s  9.40%  github.com/coreos/etcd/mvcc.(*watcherGroup).delete
```
Seems like GC is kicking in randomly. Running with `GOGC=off` gives a bit better results and here is profile

```
(pprof) top
20750ms of 28770ms total (72.12%)
Dropped 83 nodes (cum <= 143.85ms)
Showing top 10 nodes out of 72 (cum >= 610ms)
      flat  flat%   sum%        cum   cum%
    6350ms 22.07% 22.07%     6620ms 23.01%  runtime.mapaccess2_fast64
    3890ms 13.52% 35.59%     5420ms 18.84%  runtime.evacuate
    3670ms 12.76% 48.35%    10580ms 36.77%  runtime.mapassign1
    1830ms  6.36% 54.71%     2270ms  7.89%  runtime.mapdelete
    1170ms  4.07% 58.78%     1170ms  4.07%  runtime.aeshash64
     970ms  3.37% 62.15%     4680ms 16.27%  github.com/coreos/etcd/mvcc.(*watcherGroup).delete
     850ms  2.95% 65.10%    27640ms 96.07%  github.com/coreos/etcd/mvcc.BenchmarkWatchableStoreSyncedCancel
     770ms  2.68% 67.78%      770ms  2.68%  runtime.memmove
     640ms  2.22% 70.00%      640ms  2.22%  runtime.memclr
     610ms  2.12% 72.12%      610ms  2.12%  sync/atomic.AddUint32
```