@nekto0n 2016-10-12
If I lower number of watchers to 100k I consistenly get:
```
BenchmarkWatchableStoreSyncedCancel-4   	2000000000	         0.04 ns/op	       0 B/op	       0 allocs/op
```